Fix PHP module crash with inconsistent error writer function
Previously, rlogit_setmessagewriter(compile_error_capture) was
only called by the rlib_compile_infix() Zend function to set the
error message writer to the PHP specific error message collector.
Also, the GString used by the custom message writer function
was allocated and freed in the rlib_compile_infix() Zend function
but there's no API to reset the message writer to the default
and the GString pointer was left lingering, causing the crash
in the second report, which can happen when PHP FPM is using
the RLIB module.

Make creating GString *error_data explicit and call it from
the PHP module initialization function.

rlib_spool() will return the error message if the report was
not run to completion for any reason.