Support PHP 8
PHP 8 added defines for "true" and "false" which caused a compile
failure. Renamed "true" and "false" members of "struct rlib_pcode_if"
to "true_value" and "false_value", respectively, to fix this.

Replaced call_user_function_ex() with call_user_function() when
compiling for PHP 8.

Defined thread-safety related macros to empty strings when
compiling for PHP 8.

Added arginfo for public functions which is mandatory for PHP 8.